+++ src/Rules/Locale/RU/Address/PostalCode.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace LaravelExtendedValidation\Rules\Locale\RU\Address;
+
+use Illuminate\Contracts\Validation\Rule;
+use Illuminate\Support\Str;
+
+class PostalCode implements Rule
+{
+    /**
+     * @inheritDoc
+     */
+    public function passes($attribute, $value): bool
+    {
+        $postalCode = Str::of($value);
+
+        return (string) $postalCode->match('/\d{6}/') === $value;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function message(): string
+    {
+        return ':attribute does not contain a valid Russian postal code';
+    }
+}